Sort of implemented Caramel Apples
Caramel Apples are now in the game as an item that can be in the inventory. However, the actor is somehow messed up so they can't be spawned or fired from the vacpack. Attempting to fire one makes it cease to exist, and causes some other interesting issues to boot.

STATUS: Needs work.